backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
transformers>=4.57.1
torch>=2.2.0
sentencepiece>=0.2.1
protobuf>=4.21.6,<5.0.0
openai==1.3.7
google-generativeai==0.3.1
python-multipart==0.0.6
httpx==0.25.2
aiofiles==23.2.1
scikit-learn>=1.7.2
numpy>=1.24.3